Add ContactDetailsModal component and implement empty contacts service
