+++ src/components/tables/ContactsTable.jsx
@@ -9,9 +9,8 @@ const ContactsTable = ({ contacts, loading, error, selectedContactId, onSelectCo
   }
   if (error) {
     return (
-      <div className="bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded relative" role="alert">
-        <strong className="font-bold">Error: </strong>
-        <span className="block sm:inline">{error}</span>
+      <div>
+
       </div>
     );
   }
